requirements.txt
python-telegram-bot>=20,<21
requests>=2.0
pytest
flake8
